requirements.txt
vkbottle==4.4.0
tortoise-orm==0.19.3
asyncpg==0.28.0
python-dotenv==1.0.0
aiohttp==3.8.5
pydantic==1.10.12
shortuuid==1.0.11